✨ Support for writing the vhost for a user's connection string into the output secret.
This helps with the common case of plugging it into a config file or env var directly. It won't work with a cross-vhost user though, so you'll have to do that elsewhere. Really this is to work around an old bug in Keda 1.x where it doesn't understand env var interpolation (and even in 2.0, it doesn't support setting the vhost independently from the connection string).