✨ Major change: outputVhost is gone and the user secret keys have changed.
RABBIT_HOST is now RABBIT_URL. Rather than outputVhost changing that, if only one vhost is given then a RABBIT_URL_VHOST key will be created for it. Additionally RABBIT_USERNAME and RABBIT_HOSTNAME keys exist for things that don't use the URI syntax (or other silly tricks.

This would be a breaking change if anyone but me was using this.